#PAF52 - Refactored everything into new folder structure
