+++ dynatrace/cluster/user_groups.py
@@ -1,7 +1,7 @@
 #!/bin/python3
 """Cluster Group Operations"""
 import user_variables  # pylint: disable=import-error
-from dynatrace.requests import request_handler as rh
+from dynatrace.framework import request_handler as rh
 from dynatrace.tenant import management_zones as mzh
 
 MZ_USER_PERMISSONS = {
