+++ dynatrace/cluster/users.py
@@ -1,6 +1,6 @@
 """User Operations in Cluster Mangement"""
-import dynatrace.requests.request_handler as rh
-from dynatrace.exceptions import ManagedClusterOnlyException
+import dynatrace.framework.request_handler as rh
+from dynatrace.framework.exceptions import ManagedClusterOnlyException
 
 # TODO add check for is_managed
