+++ dynatrace/tenant/extensions.py
@@ -1,5 +1,5 @@
 """Operations Interacting with Dynatrace Extensions"""
-from dynatrace.requests import request_handler as rh
+from dynatrace.framework import request_handler as rh
 
 ENDPOINT = rh.TenantAPIs.EXTENSIONS
