+++ dynatrace/tenant/request_naming.py
@@ -3,7 +3,7 @@
 
 import os
 import json
-from dynatrace.requests import request_handler as rh
+from dynatrace.framework import request_handler as rh
 
 ENDPOINT = rh.TenantAPIs.REQUEST_NAMING
