+++ dynatrace/tenant/services.py
@@ -1,7 +1,7 @@
 """Module for service type entity operations"""
 
-import dynatrace.tenant.topology.shared as entity_api
-import dynatrace.requests.request_handler as rh
+import dynatrace.tenant.shared as entity_api
+import dynatrace.framework.request_handler as rh
 
 
 def get_services_tenantwide(cluster, tenant):
