+++ tests/test_maintenance_windows.py
@@ -3,8 +3,8 @@ import unittest
 import user_variables  # pylint: disable=import-error
 from tests import tooling_for_test
 from dynatrace.tenant import maintenance
-from dynatrace.requests.request_handler import TenantAPIs
-from dynatrace.exceptions import InvalidDateFormatException
+from dynatrace.framework.request_handler import TenantAPIs
+from dynatrace.framework.exceptions import InvalidDateFormatException
 
 CLUSTER = user_variables.FULL_SET["mockserver1"]
 TENANT = "tenant1"
