+++ tests/test_topology_hosts.py
@@ -4,8 +4,8 @@ Test Suite for Topology Hosts
 import unittest
 from user_variables import FULL_SET  # pylint: disable=import-error
 from tests import tooling_for_test as testtools
-from dynatrace.requests.request_handler import TenantAPIs
-from dynatrace.tenant.topology import hosts
+from dynatrace.framework.request_handler import TenantAPIs
+from dynatrace.tenant import hosts
 
 CLUSTER = FULL_SET["mockserver1"]
 TENANT = "tenant1"
